Move segmentation_labeling_app -> slapp
Updates file names, imports, and tests for master state
after merging #37
