--- tests/transforms/test_transformations.py
@@ -1,177 +1,0 @@
-import pytest
-import numpy as np
-import cv2
-import imageio
-import h5py
-import segmentation_labeling_app.transforms.transformations as transformations
-
-test_frame = np.arange(25).reshape(5, 5)
-test_video = np.stack([test_frame]*2, axis=0)
-
-
-@pytest.fixture()
-def video_fixture():
-    video = np.array(
-        [[[5, 5], [5, 5]], [[2, 2], [2, 2]],
-         [[4, 4], [4, 4]], [[0, 0], [0, 0]],
-         [[3, 2], [10, 7]], [[11, 1], [0, 5]],
-         [[8, 2], [12, 1]], [[6, 3], [2, 8]],
-         [[11, 3], [1, 17]], [[8, 3], [9, 0]]]
-    )
-    yield video
-
-
-@pytest.mark.parametrize(
-        ("array, input_fps, output_fps, random_seed, strategy, expected"),
-        [
-            (
-                # random downsample 1D array
-                np.array([1, 4, 6, 2, 3, 5, 11]),
-                7, 2, 0, 'random',
-                np.array([2, 5])),
-            (
-                # random downsample ND array
-                np.array([
-                    [1, 3], [4, 4], [6, 8], [2, 1], [3, 2],
-                    [5, 8], [11, 12]]),
-                7, 2, 0, 'random',
-                np.array([[2, 1], [5, 8]])),
-            (
-                # first downsample 1D array
-                np.array([1, 4, 6, 2, 3, 5, 11]),
-                7, 2, 0, 'first',
-                np.array([1, 3])),
-            (
-                # random downsample ND array
-                np.array([
-                    [1, 3], [4, 4], [6, 8], [2, 1], [3, 2],
-                    [5, 8], [11, 12]]),
-                7, 2, 0, 'first',
-                np.array([[1, 3], [3, 2]])),
-            (
-                # last downsample 1D array
-                np.array([1, 4, 6, 2, 3, 5, 11]),
-                7, 2, 0, 'last',
-                np.array([2, 11])),
-            (
-                # last downsample ND array
-                np.array([
-                    [1, 3], [4, 4], [6, 8], [2, 1], [3, 2],
-                    [5, 8], [11, 12]]),
-                7, 2, 0, 'last',
-                np.array([[2, 1], [11, 12]])),
-            (
-                # average downsample 1D array
-                np.array([1, 4, 6, 2, 3, 5, 11]),
-                7, 2, 0, 'average',
-                np.array([13/4, 19/3])),
-            (
-                # average downsample ND array
-                np.array([
-                    [1, 3], [4, 4], [6, 8], [2, 1], [3, 2],
-                    [5, 8], [11, 12]]),
-                7, 2, 0, 'average',
-                np.array([[13/4, 4], [19/3, 22/3]])),
-            (
-                # maximum downsample 1D array
-                np.array([1, 4, 6, 2, 3, 5, 11]),
-                7, 2, 0, 'maximum',
-                np.array([6, 11])),
-            ])
-def test_downsample(array, input_fps, output_fps, random_seed, strategy,
-                    expected):
-    array_out = transformations.downsample_array(
-            array=array,
-            input_fps=input_fps,
-            output_fps=output_fps,
-            strategy=strategy,
-            random_seed=random_seed)
-    assert np.array_equal(expected, array_out)
-
-
-@pytest.mark.parametrize(
-        ("array, input_fps, output_fps, strategy, expected"),
-        [
-            (
-                # average downsample video file with this dataset:
-                np.array([
-                    [[1, 1], [1, 1]],
-                    [[2, 2], [2, 2]],
-                    [[3, 3], [3, 3]],
-                    [[4, 4], [4, 4]],
-                    [[5, 5], [5, 5]],
-                    [[6, 6], [6, 6]],
-                    [[7, 7], [7, 7]]]),
-                7, 2, 'average',
-                np.array([
-                    [[2.5, 2.5], [2.5, 2.5]],
-                    [[6.0, 6.0], [6.0, 6.0]]])),
-                ])
-def test_video_downsample(
-        array, input_fps, output_fps, strategy, expected, tmp_path):
-
-    video_file = tmp_path / "sample_video_file.h5"
-    with h5py.File(video_file, "w") as h5f:
-        h5f.create_dataset('data', data=array)
-
-    downsampled_video = transformations.downsample_h5_video(
-            video_file,
-            input_fps,
-            output_fps,
-            strategy)
-
-    assert np.array_equal(downsampled_video, expected)
-
-
-@pytest.mark.parametrize(
-        ("array, input_fps, output_fps, random_seed, strategy, expected"),
-        [
-            (
-                # upsampling not defined
-                np.array([1, 4, 6, 2, 3, 5, 11]),
-                7, 11, 0, 'maximum',
-                np.array([6, 11])),
-            (
-                # maximum downsample ND array
-                # not defined
-                np.array([
-                    [1, 3], [4, 4], [6, 8], [2, 1], [3, 2],
-                    [5, 1234], [11, 12]]),
-                7, 2, 0, 'maximum',
-                np.array([[6, 8], [11, 12]])),
-            ])
-def test_downsample_exceptions(array, input_fps, output_fps, random_seed,
-                               strategy, expected):
-    with pytest.raises(ValueError):
-        transformations.downsample_array(
-                array=array,
-                input_fps=input_fps,
-                output_fps=output_fps,
-                strategy=strategy,
-                random_seed=random_seed)
-
-
-def test_mp4_conversion(video_fixture, tmp_path):
-    output_path = tmp_path / 'video.mp4'
-
-    fps = 30
-
-    transformations.transform_to_mp4(video=video_fixture,
-                                     output_path=output_path.as_posix(),
-                                     fps=fps),
-
-    frames = []
-    reader = imageio.get_reader(output_path.as_posix(), mode='I', fps=fps,
-                                size=(2, 2), pixelformat="gray")
-    for frame in reader:
-        gray_frame = cv2.cvtColor(frame, cv2.COLOR_RGB2GRAY)
-        frames.append(gray_frame)
-    frames = np.stack(frames)
-
-    # codec for mp4 encoding changes the video slightly between h5 and mp4
-    # and vice versa, so numbers are not exactly the same. Instead general
-    # structure of video array loaded. Cv color conversion does not rectify
-    # this error unfortunately, this has been attempted and small errors
-    # persist
-
-    assert frames.shape == video_fixture.shape
